+++ src/js/modules/components/Posts/Posts.vue
@@ -1,5 +1,6 @@
 <template>
   <div class="posts">
+    <p v-if="error">{{ error }}</p>
     <Post
       v-for="post in posts"
       :key="post.id"
@@ -12,51 +13,35 @@
 <script>
 import Post from './Post.vue';
 import Loading from '../UI/Loading.vue';
-
-const url = 'https://jsonplaceholder.typicode.com/posts';
+import { mapGetters, mapActions } from 'vuex';
 
 export default {
   components: {
     Post,
     Loading
   },
-  created() {
-    this.$store.dispatch({type: 'isLoadingPosts'});
-
-    fetch(url)
-      .then(response => response.json())
-      .then(json => {
-        this.$store.dispatch({
-          type: 'addToActivityLog',
-          message: `Data retrieved from ${url}`
-        });
-
-        setTimeout(() => {
-          this.$store.dispatch({type: 'isNotLoadingPosts'});
-          this.$store.dispatch({
-            type: 'addPostsToStore',
-            posts: json
-          });
-        }, 1000);
+  mounted() {
+    this.fetchPosts()
+      .then(response => {
+        console.log('fetchPosts success', response);
       })
       .catch(error => {
-        console.log('error', error);
-        this.$store.dispatch({
-          type: 'addToActivityLog',
-          message: `ERROR fetching data from ${url}. See log for details.`
-        });
+        console.log('fetchPosts error', error);
       });
   },
+  methods: {
+    ...mapActions([
+      'fetchPosts',
+      'addToActivityLog'
+    ])
+  },
   computed: {
-    posts() {
-      return this.$store.getters.posts;
-    },
-    postsLoaded() {
-      return this.$store.getters.postsLoaded;
-    },
-    postsLoading() {
-      return this.$store.getters.postsLoading;
-    }
+    ...mapGetters([
+      'posts',
+      'postsLoaded',
+      'postsLoading',
+      'error'
+    ])
   }
 }
 </script>