+++ src/js/store/index.js
@@ -1,5 +1,6 @@
 import Vue from 'vue';
 import Vuex from 'vuex';
+// import fetchPosts from '../Services/posts';
 
 Vue.use(Vuex);
 
@@ -20,12 +21,13 @@ export default new Vuex.Store({
     posts: null,
     postsLoaded: false,
     postsLoading: false,
-    activityLog: []
+    activityLog: [],
+    error: null
   },
   mutations: {
     addPostsToStore(state, posts) {
-      state.posts = {...posts};
-      if (!posts.length) {
+      state.posts = [...posts];
+      if (!posts) {
         state.postsLoaded = null;
         state.activityLog = addToActivityLog(state, 'Could not load posts...');
         return;
@@ -41,26 +43,43 @@ export default new Vuex.Store({
     },
     isNotLoadingPosts(state) {
       state.postsLoading = false;
+    },
+    setError(state, error) {
+      state.error = error;
     }
   },
   actions: {
-    addPostsToStore(context, payload) {
-      context.commit('addPostsToStore', payload.posts);
-    },
     addToActivityLog(context, payload) {
       context.commit('addToActivityLog', payload.message);
     },
-    isLoadingPosts(context) {
-      context.commit('isLoadingPosts');
-    },
-    isNotLoadingPosts(context) {
-      context.commit('isNotLoadingPosts');
+    fetchPosts(context) {
+      return new Promise((resolve, reject) => {
+        context.commit('setError', null);
+        context.commit('isLoadingPosts');
+
+        const url = 'https://jsonplaceholder.typicode.com/posts';
+
+        fetch(url)
+          .then(response => response.json())
+          .then(posts => {
+            context.commit('isNotLoadingPosts');
+            context.commit('addPostsToStore', posts);
+            context.commit('addToActivityLog', `Data retrieved from ${url}`);
+            resolve(posts);
+          })
+          .catch(error => {
+            context.commit('setError', error);
+            context.commit('addToActivityLog', `ERROR fetching data from ${url}. See log for details.`);
+            reject(error);
+          });
+      });
     }
   },
   getters: {
     posts: state => state.posts,
-    postsLoaded: state => state.postsLoaded,
     postsLoading: state => state.postsLoading,
+    postsLoaded: state => state.postsLoaded,
     activityLog: state => state.activityLog,
+    error: state => state.error
   },
 });